+++ Runtime/IServicePromise.cs
@@ -1,0 +1,7 @@
+﻿namespace Nonatomic.ServiceLocator
+{
+	public interface IServicePromise
+	{
+		void Fulfill(object service);
+	}
+}