+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change Log
 
+## [0.1.1-beta] - Mar 10, 2024
+- Removed unrequired interfaces from the ServiceLocator
+
+## [0.1.0-beta] - Mar 10, 2024
+- Changed the GetService method to return a promise
+- Added unit tests for the ServiceLocator
+
 ## [0.0.0-beta] - Feb 11, 2024
 - First commit